Remove the "registry upload bulk" subcommand, moving its contents up to "registry upload".
