+++ cmd/registry/cmd/upload/testdata/discovery/apigeeregistry-v1.json
