+++ cmd/registry/cmd/upload/testdata/discovery/apis.json
