+++ cmd/registry/cmd/upload/testdata/discovery/discovery-v1.json
