+++ cmd/registry/cmd/upload/testdata/openapi/apigee/registry/v1/openapi.yaml
