+++ cmd/registry/cmd/upload/testdata/openapi/petstore/2.0/swagger.yaml
