+++ cmd/registry/cmd/upload/testdata/openapi/petstore/3.0/openapi.yaml
