+++ cmd/registry/cmd/upload/testdata/protos/google/api/annotations.proto
