+++ cmd/registry/cmd/upload/testdata/protos/google/api/client.proto
