+++ cmd/registry/cmd/upload/testdata/protos/google/api/http.proto
