+++ cmd/registry/cmd/upload/testdata/protos/google/api/httpbody.proto
