+++ cmd/registry/cmd/upload/testdata/protos/google/api/resource.proto
