+++ cmd/registry/cmd/upload/testdata/protos/google/cloud/apigeeregistry/v1/BUILD.bazel
