+++ cmd/registry/cmd/upload/testdata/protos/google/cloud/apigeeregistry/v1/apigeeregistry_grpc_service_config.json
