+++ cmd/registry/cmd/upload/testdata/protos/google/cloud/apigeeregistry/v1/apigeeregistry_v1.yaml
