+++ cmd/registry/cmd/upload/testdata/protos/google/cloud/apigeeregistry/v1/provisioning_service.proto
