+++ cmd/registry/cmd/upload/testdata/protos/google/cloud/apigeeregistry/v1/registry_models.proto
