+++ cmd/registry/cmd/upload/testdata/protos/google/example/library/README.md
