+++ cmd/registry/cmd/upload/testdata/protos/google/example/library/v1/library.proto
