+++ cmd/registry/cmd/upload/testdata/protos/google/example/library/v1/library_example_gapic.yaml
