+++ cmd/registry/cmd/upload/testdata/protos/google/example/library/v1/library_grpc_service_config.json
