+++ cmd/registry/cmd/upload/testdata/protos/google/longrunning/operations.proto
