+++ cmd/registry/cmd/upload/testdata/protos/google/rpc/status.proto
