+++ cmd/registry/cmd/upload/upload.go
@@ -15,7 +15,11 @@
 package upload
 
 import (
-	"github.com/apigee/registry/cmd/registry/cmd/upload/bulk"
+	"errors"
+	"fmt"
+
+	"github.com/apigee/registry/log"
+	"github.com/apigee/registry/pkg/connection"
 	"github.com/spf13/cobra"
 )
 
@@ -25,8 +29,45 @@ func Command() *cobra.Command {
 		Short: "Upload information to the API Registry",
 	}
 
-	cmd.AddCommand(bulk.Command())
 	cmd.AddCommand(csvCommand())
+	cmd.AddCommand(discoveryCommand())
+	cmd.AddCommand(openAPICommand())
+	cmd.AddCommand(protosCommand())
+
+	cmd.PersistentFlags().String("project-id", "", "project ID to use for each upload (deprecated)")
+	cmd.PersistentFlags().String("parent", "", "parent for the upload (projects/PROJECT/locations/LOCATION)")
+	cmd.PersistentFlags().IntP("jobs", "j", 10, "number of actions to perform concurrently")
 
 	return cmd
+}
+
+func getParent(cmd *cobra.Command) (string, error) {
+	ctx := cmd.Context()
+
+	parent, err := cmd.Flags().GetString("parent")
+	if err != nil {
+		return "", fmt.Errorf("failed to get parent from flags (%s)", err)
+	}
+	projectID, err := cmd.Flags().GetString("project-id")
+	if err != nil {
+		return "", fmt.Errorf("failed to get project-id from flags (%s)", err)
+	}
+	if projectID != "" && parent != "" {
+		return "", errors.New("--project-id cannot be used with --parent")
+	}
+	if parent != "" {
+		return parent, nil
+	} else if projectID != "" {
+		log.FromContext(ctx).Warn("--project-id is deprecated, please use --parent or configure registry.project")
+		return "projects/" + projectID + "/locations/global", nil
+	}
+	c, err := connection.ActiveConfig()
+	if err != nil {
+		return "", fmt.Errorf("unable to identify parent (%s)", err)
+	}
+	parent, err = c.ProjectWithLocation()
+	if err != nil {
+		return "", fmt.Errorf("unable to identify parent: please use --parent or set registry.project in configuration (%s)", err)
+	}
+	return parent, nil
 }
